requirements.txt
altgraph       =               0.17.2
asgiref          =             3.7.2
certifi              =         2024.2.2
charset-normalizer=            3.3.2
cloudinary                 =   1.38.0
coreapi                  =     2.3.3
coreschema           =         0.0.4
distlib                  =     0.3.8
Django                   =     4.2.9
django-cloudinary-storage =    0.3.0
django-cors-headers       =    4.3.1
djangorestframework       =    3.14.0
djangorestframework-simplejwt= 5.3.1
drf-yasg                    =  1.21.7
filelock               =       3.13.1
future                   =     0.18.2
idna                    =      3.6
inflection             =       0.5.1
itypes                   =     1.2.0
Jinja2                     =   3.1.3
macholib               =       1.15.2
MarkupSafe           =         2.1.5
openapi-codec       =          1.3.2
packaging              =       23.2
pillow                   =     10.2.0
pip                         =  24.0
platformdirs            =      4.1.0
PyJWT                      =   2.8.0
python-decouple       =        3.8
pytz                       =   2023.4
PyYAML                  =      6.0.1
requests                  =    2.31.0
rest-framework-simplejwt =     0.0.2
setuptools                 =   58.0.4
six                        =   1.15.0
sqlparse               =       0.4.4
typing_extensions  =           4.9.0
uritemplate                =   4.1.1
urllib3                  =     1.26.6
virtualenv             =       20.25.0
wheel                     =    0.37.0
